Revert "-fgo-dump-spec: skip typedefs that match struct tag"
This partially reverts the following commit:

c33be99ac0c ("-fgo-dump-spec: skip typedefs that match struct tag")

It appears size_t is being treated as a struct tag and thus it isn't
emitted in the -fgo-dump-spec output. This breaks bootstrapping gccgo.

We keep the following follow-up commit so that gcc doesn't ICE later
when the original_type is later used:

7bcc83cb175 ("-fdump-go-spec: don't crash if DECL_ORIGINAL_TYPE is NULL")